refactor(tab-manager): reorganize src/lib by domain
Split flat lib/ into domain-grouped modules:
- device/ — composite IDs (branded types, create/parse) and device identity
- schema/ — table definitions, row converters, barrel with BrowserDb type
- services/ — suspend-tab helpers
- state/ — popup reactive state (browser-state, suspended-tab-state)
- utils/ — shared formatting (getDomain, getRelativeTime)

Delete dead files: epicenter/schema.ts (18-line re-export),
epicenter/index.ts (1-consumer barrel). Deduplicate getDomain across
SuspendedTabList and TabItem. Unify composite ID functions (previously
split across browser.schema.ts and device-id.ts) into single module.
