--- apps/tab-manager/src/lib/epicenter/index.ts
@@ -1,9 +1,0 @@
-export type {
-	GroupCompositeId,
-	SuspendedTab,
-	Tab,
-	TabCompositeId,
-	TabGroup,
-	Window,
-	WindowCompositeId,
-} from './browser.schema';
